+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Upcoming
+
+- `#67`: As dev, I want to standardize `require()` without `.js`.
+- `#49`: As dev/test, I want to refactor code and do jest testing. (`scp.login` to `core-remote.login`)
+
 ## `0.7.0`
 
 - `#50`: As scp user, I want to `quiet` to turn off option verification info & warning.